refactored and added new JS function updateFeedback
